# apps/web-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业级 AI Agent Harness API",
    "description": "用于企业级 AI Agent Harness 平台的任务、事件、沙箱、审计和设置 API。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "创建任务",
        "description": "创建一条新的任务记录，并写入 TASK_CREATED 事件。",
        "operationId": "create_task_api_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务列表",
        "description": "按组织查询任务列表，支持状态过滤和分页大小。",
        "operationId": "list_tasks_api_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务详情",
        "description": "返回指定任务的完整基础信息。",
        "operationId": "get_task_api_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/start": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "启动任务",
        "description": "将 CREATED 或 FAILED 任务启动为运行态。",
        "operationId": "start_task_api_tasks__task_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "取消任务",
        "description": "将当前任务状态切换为 CANCELLED，并写入 TASK_CANCELLED 事件。",
        "operationId": "cancel_task_api_tasks__task_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/resume": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "恢复任务",
        "description": "将 FAILED 或 CANCELLED 任务恢复为 RUNNING，并写入 TASK_RESUMED 事件。",
        "operationId": "resume_task_api_tasks__task_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/result": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务结果",
        "description": "返回任务状态、摘要、执行计划、产物列表和最后事件序号。",
        "operationId": "get_task_result_api_tasks__task_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/plan": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务执行计划",
        "description": "返回任务最新执行计划，并合并已落库步骤的当前状态。",
        "operationId": "get_task_plan_api_tasks__task_id__plan_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/steps": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询任务步骤",
        "description": "返回任务已执行或已派生的步骤状态列表。",
        "operationId": "list_task_steps_api_tasks__task_id__steps_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStepPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/replay": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "重放任务状态",
        "description": "根据事件流和快照重放任务状态，返回故障点和诊断摘要。",
        "operationId": "replay_task_api_tasks__task_id__replay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/model-calls": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询模型调用",
        "description": "返回当前任务关联的模型调用审计列表。",
        "operationId": "list_model_calls_api_tasks__task_id__model_calls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCallPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/tool-calls": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "查询工具调用",
        "description": "返回当前任务关联的工具调用审计列表。",
        "operationId": "list_tool_calls_api_tasks__task_id__tool_calls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/tools/execute": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "执行任务工具",
        "description": "按工具注册表和策略执行工具，并写入工具调用审计与事件流。",
        "operationId": "execute_task_tool_api_tasks__task_id__tools_execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplayRequest": {
        "properties": {
          "sequence": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence",
            "description": "重放到指定事件序号"
          }
        },
        "type": "object",
        "title": "ReplayRequest"
      },
...
      "TaskCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "任务标题"
          },
          "goal": {
            "type": "string",
            "minLength": 1,
            "title": "Goal",
            "description": "任务目标"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider",
            "description": "模型供应商"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "模型名称"
          },
          "max_runtime_seconds": {
            "type": "integer",
            "title": "Max Runtime Seconds",
            "description": "最大运行秒数",
            "default": 1800
          },
          "max_subagents": {
            "type": "integer",
            "title": "Max Subagents",
            "description": "最大子 Agent 数",
            "default": 5
          },
          "enable_sandbox": {
            "type": "boolean",
            "title": "Enable Sandbox",
            "description": "是否启用容器沙箱",
            "default": true
          },
          "enable_network": {
            "type": "boolean",
            "title": "Enable Network",
            "description": "是否启用网络访问",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "goal",
          "model_provider",
          "model_name"
        ],
        "title": "TaskCreateRequest"
      },
      "TaskPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "任务列表"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "下一页游标"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TaskPage"
      },
...
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "任务 ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "任务标题"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "description": "任务目标"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider",
            "description": "模型供应商"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "模型名称"
          },
          "max_runtime_seconds": {
            "type": "integer",
            "title": "Max Runtime Seconds",
            "description": "最大运行秒数"
          },
          "max_subagents": {
            "type": "integer",
            "title": "Max Subagents",
            "description": "最大子 Agent 数"
          },
          "enable_sandbox": {
            "type": "boolean",
            "title": "Enable Sandbox",
            "description": "是否启用容器沙箱"
          },
          "enable_network": {
            "type": "boolean",
            "title": "Enable Network",
            "description": "是否启用网络访问"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "完成时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "goal",
          "status",
          "model_provider",
          "model_name",
          "max_runtime_seconds",
          "max_subagents",
          "enable_sandbox",
          "enable_network",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}